renames(argismonitor): publish-surface rename + legacy shim + provenance docs
