+++ desktop-electrobun/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@phenotype/desktop-shell",
+  "name": "@argismonitor/desktop-shell",
   "version": "0.1.0",
   "private": true,
   "scripts": {
