+++ management-console/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "omniroute-management-console",
+  "name": "argismonitor-management-console",
   "version": "0.1.0",
   "private": true,
   "type": "module",
