+++ open-sse/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@omniroute/open-sse",
-  "version": "3.8.44",
-  "description": "Express SSE sidecar for OmniRoute — handles streaming, protocol translation, and provider orchestration",
+  "name": "@argismonitor/open-sse",
+  "version": "3.8.44-argismonitor-prep.1",
+  "description": "Express SSE sidecar for ArgisMonitor (formerly OmniRoute) — handles streaming, protocol translation, and provider orchestration",
   "type": "module",
   "main": "index.js",
   "types": "types.d.ts",
